+++ .travis.yml
@@ -1,0 +1,58 @@
+language: java
+jdk:
+  - openjdk11
+
+branches:
+  only:
+    - master
+    - develop
+    # Travis CI 서버의 Home
+cache:
+  directories:
+    - '$HOME/.m2/repository'
+    - '$HOME/.gradle'
+
+script: "./gradlew clean build"
+
+before_install:
+  - chmod +x gradlew
+
+before_deploy:
+  - zip -r nutee-back *
+  - mkdir -p deploy
+  - mv nutee-auth.zip deploy/nutee-back.zip
+
+deploy:
+  - provider: s3
+    access_key_id: $AWS_ACCESS_KEY # Travis repo settings에 설정된 값
+    secret_access_key: $AWS_SECRET_KEY # Travis repo settings에 설정된 값
+    bucket: nutee-storage # 6-3-3에서 생성한 S3 버킷
+    region: ap-northeast-2
+    skip_cleanup: true
+    acl: public_read
+    local_dir: deploy # before_deploy에서 생성한 디렉토리
+    wait-until-deployed: true
+    on:
+      repo: team-nutee/NUTEE-BACK #Github 주소
+      branch: master
+
+  - provider: codedeploy
+    access_key_id: $AWS_ACCESS_KEY # Travis repo settings에 설정된 값
+    secret_access_key: $AWS_SECRET_KEY # Travis repo settings에 설정된 값
+    bucket: nutee-storage # S3 버킷
+    key: nutee-auth.zip # S3 버킷에 저장된 nutee-back.zip 파일을 EC2로 배포
+    bundle_type: zip
+    application: nutee-back # 웹 콘솔에서 등록한 CodeDeploy 어플리케이션
+    deployment_group: nutee # 웹 콘솔에서 등록한 CodeDeploy 배포 그룹
+    region: ap-northeast-2
+    wait-until-deployed: true
+    on:
+      repo: team-nutee/NUTEE-BACK
+      branch: master
+
+# CI 실행 완료시 메일로 알람
+notifications:
+  email:
+    recipients:
+      - nutee.skhu.2020@gmail.com
+      - mf311110@gmail.com
