+++ build.gradle
@@ -41,7 +41,6 @@ dependencies {
     testImplementation('org.springframework.boot:spring-boot-starter-test') {
         exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
     }
-    testImplementation 'org.hamcrest:hamcrest-library:1.3'
 }
 
 ext {
